Refacto: - Using Mailer instead of SwiftMailer in Swift_Mailer - Renaming Swift_Mailer in SymfonyMailer
